changed include paths so they will work when installed
Include paths for all zcm core files will now work when installed in the
system. We may want to cut down the number of headers that are actually
installed, as well as the naming and organization of some of them
(zcm/cpp/zcm-cpp.hpp comes to mind).

Added a env file to be sourced by bashrc to add important things to
paths. Currently only be used to augment the CLASSPATH, but could be
used for other environment changes if necessary (seems unlikely).