changed embedded tar test to be a task instead of a post func
this makes it run more concurrently with the build and prevents it from
running during an install. Running after an install was previously a
problem because the user would run the install as sudo, so the files
generated by the post-func would be sudo priviledges only, and would
prevent the user from cleaning the build directory properly.