fixed up the static registration of transports
Static registration of transports working for share libraries regardless
of whether optimizations are enabled. Previously, they didn't work if
optimizations were turned OFF.

Additionally, it seems there is no way to force the registrations to
work statically for static libraries. This is an annoying restriction
that essentailly forces us to do an install each time we want to test
the core zcm library. We should try to find a work around for this. One
possibility would be to use a not-installed header included only in
programs that reference the static library to make a reference
to each of the transports, thereby forcing their initialization (and the
initialization of the static registers contained within them). I intend
to explore this option further after finishing the more api related
tasks.