post rebase fixes
Basically applying the changes to the commits this branch was rebased on
top of that would have been made earlier to bring them up to speed with
new features.
- moved a test into the test folder
- changed static registration of udpm transport

Also just a couple comments on cursory examination of some of the new
stuff, I'll do a more in depth review later.