Couple of fixes just to make master compile (examples) ... shame
Change of the recv_buf struct made lots of code break (including
generated code). Needed to change those member names everywhere.

Addition of the reference accepting publish function in the cpp api
cause the compiler to pick the wrong version of the overloaded function
when trying to pass a pointer. We need to have one or the other (pointer
vs reference) but it seems having both will cause build errors.

There is still a known error where zcm_stop will deadlock, but it's
easier for me to address that in my unsub branch (which hopefully won't
have to be rebased again). There is a lot of architecture fixes in there
that should make this type of stuff easier to detect and fix.