changed my mind about changing the java implementation of sub/unsub
Basically the port from LCM left the java code in a weird state because
lcm doesn't use JNI so they had to rewrite a LOT of the backbone.
Changing the sub/unsub api in java is going to be a large rewrite that
isn't really worth it right now.

Note that this means that java can't really ever turn off the underlying
subscriptions, it just receives the messages and decides not to handle
them. This will result in performance costs when you unsubscribe from
things in java because you will technically still be receiving the
messages, just doing nothing with them.