Give users control over the march arg passed to the compiler
Before this commit, zcm would always build with -march=native which
instructs the compiler to detect the optimal cpu architecture from the
host. However, this can result in binaries that will not be able to run
on other cpu architectures (see the gcc man page for more information).
In order to make it so that zcm can be built in a more flexible manner,
this commit gives users the ability to set the march command.